+++ packages/frontend/src/components/CreateAccount/EmailForm.jsx
@@ -1,0 +1,13 @@
+// @flow
+import React from 'react';
+import { FormControl, Input, InputLabel } from '@material-ui/core';
+import type { Node } from 'react';
+
+const EmailForm = (): Node => (
+  <FormControl margin="normal" required fullWidth>
+    <InputLabel htmlFor="email">Email Address</InputLabel>
+    <Input id="email" name="email" autoComplete="email" />
+  </FormControl>
+);
+
+export default EmailForm;
